fix!: @default not rendered on required properties (#144)
BREAKING CHANGE: The @default doctag is now rendered by default for required properties. Use `outputFileOptions.defaultTagsForRequiredProps` to restore the previous behavior.

The @default tag is used by `projen new` to populate required options that are not provided via the cli. In hindsight this should not have been a responsibility for the renderer. The option is now deprecated and will be removed once an easy way to remove the tag for all required properties has been added.

Fixes #142