docs: add more examples, suggest using `devDeps` in .projenrc to install this package (#231)
I had some trouble figuring out how to use this, because I'm unfamiliar with jsii's spec API. So I've added some more examples for complex types, for others who have the same problem.

There are some other things in the examples, and a suggestion to use projen to install this package instead of directly using npm.