feat: allow full Docs for top level of ProjenStruct (#233)
This is what I wanted in that issue I opened. I made the new `docs` take precedence over the old `description`. Hopefully this is an acceptable change.

While I'm here again, I changed the example of a long doc string that I added in my last PR, to be a bit better syntax.

Fixes #230